WEA-22 Seperate hourly data to seperate class from inner class
